Move ClusterConfig to internal types package
