+++ src/k8s/pkg/k8sd/database/cluster_config_test.go
@@ -5,7 +5,8 @@ import (
 	"database/sql"
 	"testing"
 
-	"github.com/canonical/k8s/pkg/k8sd/database/clusterconfigs"
+	"github.com/canonical/k8s/pkg/k8sd/database"
+	"github.com/canonical/k8s/pkg/k8sd/types"
 	. "github.com/onsi/gomega"
 )
 
@@ -13,8 +14,8 @@ func TestClusterConfig(t *testing.T) {
 	WithDB(t, func(ctx context.Context, d DB) {
 		t.Run("Set", func(t *testing.T) {
 			g := NewWithT(t)
-			expectedClusterConfig := clusterconfigs.ClusterConfig{
-				Certificates: clusterconfigs.Certificates{
+			expectedClusterConfig := types.ClusterConfig{
+				Certificates: types.Certificates{
 					CACert: "CA CERT DATA",
 					CAKey:  "CA KEY DATA",
 				},
@@ -22,7 +23,7 @@ func TestClusterConfig(t *testing.T) {
 
 			// Write some config to the database
 			err := d.Transaction(ctx, func(ctx context.Context, tx *sql.Tx) error {
-				err := clusterconfigs.SetClusterConfig(context.Background(), tx, expectedClusterConfig)
+				err := database.SetClusterConfig(context.Background(), tx, expectedClusterConfig)
 				g.Expect(err).To(BeNil())
 				return nil
 			})
@@ -30,7 +31,7 @@ func TestClusterConfig(t *testing.T) {
 
 			// Retrieve it and map it to the struct
 			err = d.Transaction(ctx, func(ctx context.Context, tx *sql.Tx) error {
-				clusterConfig, err := clusterconfigs.GetClusterConfig(ctx, tx)
+				clusterConfig, err := database.GetClusterConfig(ctx, tx)
 				g.Expect(err).To(BeNil())
 				g.Expect(clusterConfig).To(Equal(expectedClusterConfig))
 				return nil
@@ -41,16 +42,16 @@ func TestClusterConfig(t *testing.T) {
 		t.Run("CannotUpdateCA", func(t *testing.T) {
 			// TODO(neoaggelos): extend this test for all fields that cannot be updated
 			g := NewWithT(t)
-			expectedClusterConfig := clusterconfigs.ClusterConfig{
-				Certificates: clusterconfigs.Certificates{
+			expectedClusterConfig := types.ClusterConfig{
+				Certificates: types.Certificates{
 					CACert: "CA CERT DATA",
 					CAKey:  "CA KEY DATA",
 				},
 			}
 
 			err := d.Transaction(ctx, func(ctx context.Context, tx *sql.Tx) error {
-				err := clusterconfigs.SetClusterConfig(context.Background(), tx, clusterconfigs.ClusterConfig{
-					Certificates: clusterconfigs.Certificates{
+				err := database.SetClusterConfig(context.Background(), tx, types.ClusterConfig{
+					Certificates: types.Certificates{
 						CACert: "CA CERT NEW DATA",
 					},
 				})
@@ -60,7 +61,7 @@ func TestClusterConfig(t *testing.T) {
 			g.Expect(err).To(HaveOccurred())
 
 			err = d.Transaction(ctx, func(ctx context.Context, tx *sql.Tx) error {
-				clusterConfig, err := clusterconfigs.GetClusterConfig(ctx, tx)
+				clusterConfig, err := database.GetClusterConfig(ctx, tx)
 				g.Expect(err).To(BeNil())
 				g.Expect(clusterConfig).To(Equal(expectedClusterConfig))
 				return nil
@@ -69,33 +70,32 @@ func TestClusterConfig(t *testing.T) {
 		})
 
 		t.Run("Update", func(t *testing.T) {
-			// TODO(neoaggelos): extend this test for all fields that can be updated
 			g := NewWithT(t)
-			expectedClusterConfig := clusterconfigs.ClusterConfig{
-				Certificates: clusterconfigs.Certificates{
+			expectedClusterConfig := types.ClusterConfig{
+				Certificates: types.Certificates{
 					CACert:        "CA CERT DATA",
 					CAKey:         "CA KEY DATA",
 					K8sDqliteCert: "CERT DATA",
 					K8sDqliteKey:  "KEY DATA",
 				},
-				Kubelet: clusterconfigs.Kubelet{
+				Kubelet: types.Kubelet{
 					ClusterDNS: "10.152.183.10",
 				},
-				APIServer: clusterconfigs.APIServer{
+				APIServer: types.APIServer{
 					ServiceAccountKey: "SA KEY DATA",
 				},
 			}
 
 			err := d.Transaction(ctx, func(ctx context.Context, tx *sql.Tx) error {
-				err := clusterconfigs.SetClusterConfig(context.Background(), tx, clusterconfigs.ClusterConfig{
-					Kubelet: clusterconfigs.Kubelet{
+				err := database.SetClusterConfig(context.Background(), tx, types.ClusterConfig{
+					Kubelet: types.Kubelet{
 						ClusterDNS: "10.152.183.10",
 					},
-					Certificates: clusterconfigs.Certificates{
+					Certificates: types.Certificates{
 						K8sDqliteCert: "CERT DATA",
 						K8sDqliteKey:  "KEY DATA",
 					},
-					APIServer: clusterconfigs.APIServer{
+					APIServer: types.APIServer{
 						ServiceAccountKey: "SA KEY DATA",
 					},
 				})
@@ -105,7 +105,7 @@ func TestClusterConfig(t *testing.T) {
 			g.Expect(err).To(BeNil())
 
 			err = d.Transaction(ctx, func(ctx context.Context, tx *sql.Tx) error {
-				clusterConfig, err := clusterconfigs.GetClusterConfig(ctx, tx)
+				clusterConfig, err := database.GetClusterConfig(ctx, tx)
 				g.Expect(err).To(BeNil())
 				g.Expect(clusterConfig).To(Equal(expectedClusterConfig))
 				return nil
